+++ src/config/app.exceptions.ts
@@ -1,0 +1,9 @@
+import {
+  InternalServerErrorException,
+  UnauthorizedException,
+} from '@nestjs/common';
+
+export const appExceptions = {
+  unauthorized: new UnauthorizedException(),
+  unknown: new InternalServerErrorException(),
+};