+++ src/posts/config/posts.exceptions.ts
@@ -1,11 +1,14 @@
 import {
   BadRequestException,
+  ConflictException,
   ForbiddenException,
   HttpException,
   HttpStatus,
   InternalServerErrorException,
   NotFoundException,
 } from '@nestjs/common';
+import { appExceptions } from 'src/config/app.exceptions';
+import { validationException } from 'src/validation/validation.pipe';
 
 export const postsExceptions = {
   invalidThreadId: new BadRequestException('Must provide a valid thread id.'),
@@ -19,6 +22,18 @@ export const postsExceptions = {
     'Create or edit request failed due to an unknown reason.',
   ),
   quote: {
+    unauthorized: appExceptions.unauthorized,
+    validationFailed: validationException,
     notFound: new NotFoundException(),
+    unknown: appExceptions.unknown,
+  },
+  report: {
+    unauthorized: appExceptions.unauthorized,
+    validationFailed: validationException,
+    notFound: new NotFoundException(),
+    alreadyReported: new ConflictException(
+      'This post has already been reported.',
+    ),
+    unknown: appExceptions.unknown,
   },
 };